crisis_requirements.txt
accelerate==0.26.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attr==0.3.2
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
dashscope==1.20.11
datasets==3.2.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
faiss-gpu==1.7.2
fake-useragent==2.0.1
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.3
idna==3.10
importlib_resources==6.4.5
jiayan==0.0.21
jieba==0.42.1
Jinja2==3.1.3
jiter==0.6.1
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
openai==1.52.0
packaging==24.2
pandas==2.2.3
pillow==10.2.0
propcache==0.2.0
psutil==6.1.0
pyarrow==18.1.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
pyparsing==3.2.0
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.3.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.5
transformers==4.47.0
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
xxhash==3.5.0
yarl==1.15.5
zhipuai==2.1.5.20241204
zipp==3.21.0
